pyproject.toml
[tool.poetry]
name = "modular"
version = "0.1.0"
description = ""
authors = ["Stax124 <tamoncz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "2.0.0."
discord-pretty-help = "^1.3.4"
DiscordUtils = "^1.3.4"
sqlmodel = "^0.0.6"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.3"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
coloredlogs = "^15.0.1"
requests = "^2.28.1"
termcolor = "^1.1.0"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.3"

[tool.poe.tasks]

    [tool.poe.tasks.build]
    help= "Build docker container"
    cmd = "docker build -t modular ."

    [tool.poe.tasks.start]
    help= "Run the application"
    cmd = "python main.py -l DEBUG --enable-rce"

    [tool.poe.tasks.format]
    help= "Format the code"
    cmd = "black ."

    [tool.poe.tasks.web]
    help= "Run the web server"
    cmd = "uvicorn web:app --reload"

    [tool.poe.tasks.export]
    help= "Export the requirements"
    cmd = "poetry export -f requirements.txt --without-hashes"

    [tool.poe.tasks.revision]
    help= "Generate new alembic revision"
    cmd = "alembic revision --autogenerate"

    [tool.poe.tasks.migrate]
    help= "Migrate the database"
    cmd = "alembic upgrade head"

    [tool.poe.tasks.test]
    help= "Test the project"
    cmd = "python -m pytest --ignore=plugins"

    [tool.poe.tasks.doxygen]
    help= "Generate doxygen documentation"
    cmd = "doxygen"

    [tool.poe.tasks.docs]
    help= "Generate documentation"
    cmd = "doxybook2 --input doxygen/xml --output docs --config doxybook2.json"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
